tests/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno==1.7.4",
    "anthropic>=0.57.1",
    "duckduckgo-search>=8.1.1",
    "mcp>=1.11.0",
    "mcp-contextforge-gateway>=0.3.1",
    "openai>=1.96.0",
    "python-dotenv>=1.1.1",
]
